isys package lib2600 b0 for scci_master_20151228_1800

+++ drivers/media/pci/intel-ipu4/ipu4_bxtB0/lib2600/pkg_dir/interface/ia_css_pkg_dir.h
@@ -1,15 +1,15 @@
 /**
 * Support for Intel Camera Imaging ISP subsystem.
-* Copyright (c) 2010 - 2015, Intel Corporation.
-* 
-* This program is free software; you can redistribute it and/or modify it
-* under the terms and conditions of the GNU General Public License,
-* version 2, as published by the Free Software Foundation.
-* 
-* This program is distributed in the hope it will be useful, but WITHOUT
-* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-* FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
-* more details.
+ * Copyright (c) 2010 - 2015, Intel Corporation.
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms and conditions of the GNU General Public License,
+ * version 2, as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
 */
 
 #ifndef _IA_CSS_PKG_DIR_H_
@@ -38,7 +38,7 @@ uint32_t ia_css_pkg_dir_get_num_entries(
 );
 
 IA_CSS_PKG_DIR_STORAGE_CLASS_H
-size_t ia_css_pkg_dir_get_size_in_bytes(
+uint32_t ia_css_pkg_dir_get_size_in_bytes(
 	const ia_css_pkg_dir_entry_t *pkg_dir_header
 );
